+++ packages/@atjson/source-gdocs-paste/test/converter-test.ts
@@ -26,7 +26,6 @@ describe('@atjson/source-gdocs-paste', () => {
 
   it('correctly converts headings', () => {
     let headings = atjson.where(a => a.type === 'heading');
-    console.log(headings.toJSON());
     expect(headings.length).toEqual(4);
     expect(headings.map(h => h.attributes.level)).toEqual([1, 2, 100, 101]);
   });
@@ -34,6 +33,19 @@ describe('@atjson/source-gdocs-paste', () => {
   it('correctly converts lists', () => {
     let lists = atjson.where(a => a.type === 'list');
     expect(lists.length).toEqual(2);
+  });
+
+  it('adds parse tokens around newlines separating list-items', () => {
+    let lists = atjson.where(a => a.type === 'list').as('list');
+    let allParseTokens = atjson.where(a => a.type === 'parse-token').as('parseTokens');
+
+    lists.outerJoin(
+        allParseTokens,
+        (l, r) => l.start <= r.start && l.end >= r.end
+      ).forEach(({ list, parseTokens }) => {
+        let newlines = atjson.match(/\n/g, list.start, list.end);
+        expect(newlines.map(match => [ match.start, match.end ])).toEqual(parseTokens.map(a => [ a.start, a.end ]));
+      });
   });
 
   it('correctly converts numbered lists', () => {
@@ -79,15 +91,19 @@ describe('@atjson/source-gdocs-paste paragraphs', () => {
   const LINEBREAKS = [
     [ 21, 22],
     [ 70, 71 ],
+    [ 213, 214 ],
     [ 249, 250 ],
-    [ 370, 371 ]
+    [ 370, 371 ],
+    [ 486, 487 ],
+    [ 520, 521 ],
+    [ 538, 539 ]
   ].map(([start, end]) => {
     return {
       start,
       end,
       type: '-offset-line-break',
       attributes: {},
-      id: 'Any<id>'
+      id: expect.anything()
     };
   });
 
@@ -105,7 +121,7 @@ describe('@atjson/source-gdocs-paste paragraphs', () => {
       end,
       type: '-offset-paragraph',
       attributes: {},
-      id: 'Any<id>'
+      id: expect.anything()
     };
   });
 
@@ -118,8 +134,8 @@ describe('@atjson/source-gdocs-paste paragraphs', () => {
       end,
       type: '-offset-list',
       attributes: { '-offset-type': 'numbered' },
-      id: 'Any<id>'
-    }
+      id: expect.anything()
+    };
   });
 
   const LIST_ITEMS = [
@@ -133,7 +149,7 @@ describe('@atjson/source-gdocs-paste paragraphs', () => {
       end,
       type: '-offset-list-item',
       attributes: {},
-      id: 'Any<id>'
+      id: expect.anything()
     };
   });
 
@@ -156,7 +172,7 @@ describe('@atjson/source-gdocs-paste paragraphs', () => {
         (l, r) => r.end <= l.start
       );
 
-    expect(listsAndParagraphs.toJSON()[0]).toEqual({
+    expect(listsAndParagraphs.toJSON()[0]).toMatchObject({
       list: LISTS[0],
       paragraphs: PARAGRAPHS.slice(0, 3)
     });
@@ -178,7 +194,7 @@ describe('@atjson/source-gdocs-paste paragraphs', () => {
   it('created linebreaks inside list-items', () => {
     let linebreaks = atjson.where({ type: '-offset-line-break' }).as('linebreak');
 
-    expect(linebreaks.toJSON()).toEqual(LINEBREAKS);
+    expect(linebreaks.toJSON()).toMatchObject(LINEBREAKS);
 
     let linebreaksInLists = linebreaks
       .join(
@@ -191,13 +207,13 @@ describe('@atjson/source-gdocs-paste paragraphs', () => {
 
     // No linebreaks in a list outside of a list-item
     expect(linebreaksInLists.where(join => join.lists.length > 0 && join['list-items'].length === 0).toJSON()).toHaveLength(0);
-    expect(linebreaksInLists.toJSON()).toEqual([
+    expect(linebreaksInLists.toJSON()).toMatchObject([
       {
-        linebreak: LINEBREAKS[2],
+        linebreak: LINEBREAKS[3],
         lists: [ LISTS[0] ],
         'list-items': [ LIST_ITEMS[0] ]
       }, {
-        linebreak: LINEBREAKS[3],
+        linebreak: LINEBREAKS[4],
         lists: [ LISTS[0] ],
         'list-items': [ LIST_ITEMS[1] ]
       }
@@ -207,7 +223,7 @@ describe('@atjson/source-gdocs-paste paragraphs', () => {
   it('created paragraphs inside list-items', () => {
     let paragraphs = atjson.where({ type: '-offset-paragraph' }).as('paragraph');
 
-    expect(paragraphs.toJSON()).toEqual(PARAGRAPHS);
+    expect(paragraphs.toJSON()).toMatchObject(PARAGRAPHS);
 
     let paragraphsInLists = paragraphs
       .join(
@@ -220,7 +236,7 @@ describe('@atjson/source-gdocs-paste paragraphs', () => {
 
     // No paragraphs in a list outside of a list-item
     expect(paragraphsInLists.where(join => join.lists.length > 0 && join['list-items'].length === 0).toJSON()).toHaveLength(0);
-    expect(paragraphsInLists.toJSON()).toEqual([
+    expect(paragraphsInLists.toJSON()).toMatchObject([
       {
         paragraph: PARAGRAPHS[3],
         lists: [ LISTS[0] ],
@@ -245,13 +261,14 @@ describe('@atjson/source-gdocs-paste paragraphs', () => {
     let ocrs = atjson.match(/\uFFFC/g, LISTS[1].start, LISTS[1].end);
     expect(ocrs).toEqual([{
       start: LISTS[1].end - 1,
-      end: LISTS[1].end
+      end: LISTS[1].end,
+      matches: expect.anything()
     }]);
 
     let lists = atjson.where({ type: '-offset-list' });
-    expect(lists.toJSON()).toEqual(LISTS);
+    expect(lists.toJSON()).toMatchObject(LISTS);
 
     let listItems = atjson.where({ type: '-offset-list-item' });
-    expect(listItems.toJSON()).toEqual(LIST_ITEMS);
+    expect(listItems.toJSON()).toMatchObject(LIST_ITEMS);
   });
 });