+++ packages/@atjson/document/CHANGELOG.md
@@ -5,7 +5,9 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 ## [0.13.5](https://github.com/CondeNast-Copilot/atjson/compare/@atjson/document@0.13.4...@atjson/document@0.13.5) (2019-01-07)
 
-**Note:** Version bump only for package @atjson/document
+### ✨ New Features
+
+* ✨ Add query methods `document.match`, `namedCollection.outerJoin`, `namedCollection.forEach`, `namedCollection.reduce`, `join.outerJoin`, `join.where`, `join.forEach`